it works

+++ src/main/java/org/damm/entity/Category.java
@@ -20,7 +20,7 @@ public class Category {
 
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Long idCategory;
+	private Integer idCategory;
 
 	@Column(nullable = false, unique = true)
 	private String name;
